requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-qr-code==4.1.0
djangorestframework==3.15.1
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.7.2
pydantic_core==2.18.3
rsa==4.9
segno==1.6.1
sqlparse==0.5.0
Telethon==1.35.0
typing_extensions==4.12.1
tzdata==2024.1
